devide main into botgo
